Updated readme, removed product photo name from response to FE, changed in UserAddressEntity type of column mobilePhone from tinyint to int

+++ src/user/entities/user-address.entity.ts
@@ -39,7 +39,7 @@ export class UserAddressEntity
   country: string;
 
   @Column({
-    type: 'tinyint',
+    type: 'int',
   })
   mobilePhone: number;
